+++ views/layouts/main.handlebars
@@ -114,15 +114,7 @@
       </div>
     </div>
   </footer>
-      
-  {{#if logged_in}}
-  <script src="../../public/js/logout.js"></script>
-  {{/if}}
-  <script src="../../public/js/login.js"></script>
-  {{!-- <script src="../public/js/logout.js">
-  </script> --}}
-  <script src="../public/js/search.js">
-  </script>
+
 
   </body>
 </html>
