+++ views/login.handlebars
@@ -16,31 +16,30 @@
     </form>
   </div>
 </div>
-<div class="row">
-    <div class="col-md-6">
+  <div class="col-md-6">
     <h2>Signup</h2>
-    <form class="form signup-form" id="signup-form"> 
-      <div class="form-group">
-        <label for="name-signup">Name:</label>
-        <input class="form-input" type="text" id="name-signup" placeholder="Enter your name" />
-      </div>
-      <div class="form-group">
-        <label for="email-signup">Email:</label>
-        <input class="form-input" type="text" id="email-signup" placeholder="Enter your email" />
-      </div>
-      <div class="form-group">
-        <label for="password-signup">Password:</label>
-        <input class="form-input" type="password" id="password-signup" placeholder="Create a password" />
-      </div>
-      <div class="form-group">
-        <button class="btn btn-primary" type="submit">Signup</button>
-      </div>
-    </form>
+      <form class="form signup-form" id="signup-form"> 
+        <div class="form-group">
+          <label for="name-signup">Name:</label>
+          <input class="form-input" type="text" id="name-signup" placeholder="Enter your name" />
+        </div>
+        <div class="form-group">
+          <label for="email-signup">Email:</label>
+          <input class="form-input" type="text" id="email-signup" placeholder="Enter your email" />
+        </div>
+        <div class="form-group">
+          <label for="password-signup">Password:</label>
+          <input class="form-input" type="password" id="password-signup" placeholder="Create a password" />
+        </div>
+        <div class="form-group">
+          <button class="btn btn-primary" type="submit">Signup</button>
+        </div>
+      </form>
   </div>
 </div>
 
-<script src="../public/js/signup.js"></script>
-<script src="../public/login.js"></script>
+<script src="/js/login.js"></script>
+
 
 {{!-- Login and Sign up page from Ann's Homework --}}
