+++ views/profile.handlebars
@@ -1,0 +1,68 @@
+<div class="row">
+  <div class="col-auto">
+    <h2>Welcome, {{name}}!</h2>
+  </div>
+</div>
+
+<div class="row mt-4">
+  <div class="col-md-6">
+    <h3>Add Your Recipe!</h3>
+    <form class="form new-recipe-form">
+      <div class="form-group">
+        <label for="name-recipe">Recipe Name:</label>
+        <input
+          class="form-input"
+          type="text"
+          id="name-recipe"
+          placeholder="Enter your recipe"
+        />
+      </div>
+
+      <div class="form-group">
+        <label for="ingredients-recipe">Ingredients:</label>
+        <input
+          class="form-input"
+          type="text"
+          id="ingredients-recipe"
+          placeholder="Enter your ingredients"
+        />
+      </div>
+
+      <div class="form-group">
+        <label for="instructions-recipe">Instructions:</label>
+        <input
+          class="form-input"
+          type="text"
+          id="instructions-recipe"
+          placeholder="Enter your instructions"
+        />
+      </div>
+
+      <div class="form-group">
+        <button class="btn btn-primary" type="submit">Add Recipe</button>
+      </div>
+    </form>
+  </div>
+
+  {{#if recipes.length}}
+    <div class="col-md-6 recipe-list">
+      <h3>Your Recipes</h3>
+
+      {{#each recipes as |recipe|}}
+        <div class="row mb-2">
+          <div class="col-md-8">
+            <h4><a href="/recipe/{{recipe.id}}">{{recipe.name}}</a></h4>
+          </div>
+          <div class="col-md-4">
+            <button
+              class="btn btn-sm btn-danger"
+              data-id="{{recipe.id}}"
+            >DELETE</button>
+          </div>
+        </div>
+      {{/each}}
+    </div>
+  {{/if}}
+</div>
+
+<script src="/js/add-recipe.js"></script>
