+++ views/recipe.handlebars
@@ -1,38 +1,8 @@
-<div class="row">
-    <div class="col-md-6">
-        <h2>Add Your Recipe!</h2>
-        < class="form" recipe-form">
-            <div class="form-group">
-                <label for="name-recipe">Name:</label>
-                <input class="form-input" type="text" id="name-recipe" placeholder="Enter your recipe" />
-            </div>
-            <div class="form-group">
-                <label for="ingredients-recipe">Ingredients:</label>
-                <input class="form-input"> type="text" id="ingredients-recipe" placeholder="Enter your ingredients" />
-            </div>
-
-            <div class="form-group">
-                <label for="instructions-recipe">Instructions:</label>
-                <input class="form-input"> type="text" id="instructions-recipe" placeholder="Enter your instructions" />
-            </div>
-            <div class="form-group">
-                <button class="btn btn-primary" type="submit">Add Recipe</button>
-            </div>
-        </form>
-        <div class="column" id="recipesbody">
-            {{{#each recipes}}}
-            <div class="card">
-                <div class="card-body">
-                    <h5 class="card-title
-                    ">{{name}}</h5>
-                    <p class="card-text">{{ingredients}}</p>
-                    <p class="card-text">{{instructions}}</p>
-                    <a href="#" class="btn btn-primary">View Comments</a>
-                    {{{/each}}}
-
-        </div>
-    </div>
+<div class="card">
+  <div class="card-body">
+    <h5 class="card-title">{{name}}</h5>
+    <p class="card-text">{{ingredients}}</p>
+    <p class="card-text">{{instructions}}</p>
+    <a href="#" class="btn btn-primary">View Comments</a>
+  </div>
 </div>
-
-<script src="../public/js/add-recipe.js"></script>
-
